+++ client/src/components/Comment/CreateComment.tsx
@@ -3,7 +3,6 @@ import {
   Textarea,
   FormControl,
   FormLabel,
-  Box,
   VisuallyHidden,
   CardFooter,
   Flex,
