+++ client/src/pages/CreateArticlePage.tsx
@@ -1,5 +1,5 @@
 import { Box, Container, Heading } from "@chakra-ui/react";
-import CreateArticle from "../components/Article/Actions/CreateArticle";
+import CreateArticle from "../components/Article/Content/CreateArticleCard";
 import { blue_color } from "../assets/customColors";
 
 export default function CreateArticlePage(): JSX.Element {
